+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.1] - 2023-01.10
+## [0.5.0] - 2023-01-17
+### Added
+- Added more endpoints for updating relations
+- Added rudimentary check to make sure user is the one updating field
+- Added checks for some relations to make sure we're not removing the last item
+- Added more exclusions on response models
+### Changed
+- Use Tortoise ORM instead of SQLAlchemy
+- Refactor everything to use Tortoise ORM
+- Various style changes
+
+## [0.4.1] - 2023-01-10
 ### Added
 - Added register, login endpoints
 - Added OAuth support